Move is active factory into new separated file
