+++ ilc/client.js
@@ -4,6 +4,7 @@ import Router from './common/router/ClientRouter';
 import setupErrorHandlers from './client/errorHandler/setupErrorHandlers';
 import {fragmentErrorHandlerFactory, crashIlc} from './client/errorHandler/fragmentErrorHandlerFactory';
 import handlePageTransaction, {slotWillBe} from './client/handlePageTransaction';
+import isActiveFactory from './client/isActiveFactory';
 import initSpaConfig from './client/initSpaConfig';
 import setupPerformanceMonitoring from './client/performance';
 import selectSlotsToRegister from './client/selectSlotsToRegister';
@@ -51,7 +52,14 @@ selectSlotsToRegister([...registryConf.routes, registryConf.specialRoutes['404']
                 return Promise.all(waitTill)
                     .then(v => v[0].mainSpa !== undefined ? v[0].mainSpa(appConf.initProps || {}) : v[0]);
             },
-            isActiveFactory(appName, slotName),
+            isActiveFactory({
+                singleSpa,
+                router,
+                handlePageTransaction,
+                slotWillBe,
+                appName,
+                slotName,
+            }),
             {
                 domElementGetter: () => getSlotElement(slotName),
                 getCurrentPathProps: () => router.getCurrentRouteProps(appName, slotName),
@@ -63,52 +71,6 @@ selectSlotsToRegister([...registryConf.routes, registryConf.specialRoutes['404']
     });
 });
 
-function isActiveFactory(appName, slotName) {
-    let reload = false;
-
-    return () => {
-        const checkActivity = (route) => Object.entries(route.slots).some(([
-            currentSlotName,
-            slot
-        ]) => slot.appName === appName && currentSlotName === slotName);
-
-        let isActive = checkActivity(router.getCurrentRoute());
-        const wasActive = checkActivity(router.getPrevRoute());
-
-        let willBe = slotWillBe.default;
-        !wasActive && isActive && (willBe = slotWillBe.rendered);
-        wasActive && !isActive && (willBe = slotWillBe.removed);
-
-        if (isActive && wasActive && reload === false) {
-            const oldProps = router.getPrevRouteProps(appName, slotName);
-            const currProps = router.getCurrentRouteProps(appName, slotName);
-
-            if (JSON.stringify(oldProps) !== JSON.stringify(currProps)) {
-                window.addEventListener('single-spa:app-change', function singleSpaAppChange() {
-                    window.removeEventListener('single-spa:app-change', singleSpaAppChange);
-                    //TODO: need to consider addition of the new update() hook to the adapter. So it will be called instead of re-mount, if available.
-                    console.log(`ILC: Triggering app re-mount for ${appName} due to changed props.`);
-
-                    reload = true;
-
-                    singleSpa.triggerAppChange();
-                });
-
-                isActive = false;
-                willBe = slotWillBe.rerendered;
-            }
-        }
-
-        if (window.ilcConfig && window.ilcConfig.tmplSpinner) {
-            handlePageTransaction(slotName, willBe);
-        }
-
-        reload = false;
-
-        return isActive;
-    }
-}
-
 setupErrorHandlers(registryConf, router.getCurrentRoute);
 setupPerformanceMonitoring(router.getCurrentRoute);
 
@@ -117,4 +79,4 @@ singleSpa.setMountMaxTime(5000, false);
 singleSpa.setUnmountMaxTime(3000, false);
 singleSpa.setUnloadMaxTime(3000, false);
 
-singleSpa.start({ urlRerouteOnly: true });
+singleSpa.start({urlRerouteOnly: true});
